pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "graphsignal"
version = "0.0.1"
description = "Graph-based financial signal flagging engine (research prototype)"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "pandas>=2.2",
    "pyarrow>=15",
    "yfinance>=0.2.40",
    "requests>=2.32",
    "tenacity>=8.2",
    "pydantic>=2.6",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[tool.setuptools.packages.find]
include = ["graphsignal*"]
exclude = ["tests*", "data*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
